Refactor polyfill: split into different files & add comments.
