+++ libyul/backends/wasm/polyfill/eei.h
@@ -1,0 +1,524 @@
+/*
+	This file is part of solidity.
+
+	solidity is free software: you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation, either version 3 of the License, or
+	(at your option) any later version.
+
+	solidity is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with solidity.  If not, see <http://www.gnu.org/licenses/>.
+*/
+// SPDX-License-Identifier: GPL-3.0
+/**
+ * This file includes logical functions for the wasm polyfill.
+ */
+
+#pragma once
+
+#include <string>
+
+namespace solidity::yul::wasm::polyfill
+{
+
+static std::string const eei{R"(
+// solidity::yul::wasm::polyfill::eei
+
+// address of the current contract / execution context
+function address() -> z1, z2, z3, z4 {
+	// eth.getAddress: Gets address of currently executing account and stores it
+	// in memory at the given offset. This address is encoded as a 160 bit number,
+	// represented as a 20 bytes long little endian unsigned integer in memory.
+	eth.getAddress(4:i32)
+
+	// mload_address takes the little endian unsigned integer from memory and converts
+	// it to an big endian unsigned integer (evm is big endian).
+	z2, z3, z4 := mload_address(0:i32)
+}
+
+//  balance(a): wei balance at address a
+function balance(x1, x2, x3, x4) -> z1, z2, z3, z4 {
+	// address x is a big endian encoded unsigned integer.
+	// mstore_address(pos, address) will write that address as a little endian
+	// unsigned integer at offset 0.
+	mstore_address(0:i32, x1, x2, x3, x4)
+
+	// getExternalBalance(addressOffset, resultOffset): Gets balance of the
+	// given account and loads it into memory at the given offset.
+	// address need to be encoded as a little endian unsigned integer.
+	// balance is of type u128 represented as a 16 bytes long
+	// little endian unsigned integer in memory.
+	eth.getExternalBalance(12:i32, 32:i32)
+
+	z3 := i64.load(40:i32)
+	z4 := i64.load(32:i32)
+}
+
+// equivalent to balance(address()), but cheaper
+function selfbalance() -> z1, z2, z3, z4 {
+	// TODO: not part of current Ewasm spec
+	unreachable()
+}
+
+// ID of the executing chain (EIP 1344)
+function chainid() -> z1, z2, z3, z4 {
+	// TODO: not part of current Ewasm spec
+	unreachable()
+}
+
+// transaction sender
+function origin() -> z1, z2, z3, z4 {
+	eth.getTxOrigin(4:i32)
+	z2, z3, z4 := mload_address(0:i32)
+}
+
+// call sender (excluding delegatecall)
+function caller() -> z1, z2, z3, z4 {
+	eth.getCaller(4:i32)
+	z2, z3, z4 := mload_address(0:i32)
+}
+
+// wei sent together with the current call
+function callvalue() -> z1, z2, z3, z4 {
+	eth.getCallValue(0:i32)
+	z1, z2, z3, z4 := mload_internal(0:i32)
+}
+
+// call data starting from position x (32 bytes)
+function calldataload(x1, x2, x3, x4) -> z1, z2, z3, z4 {
+	mstore_internal(0:i32, 0, 0, 0, 0)
+	calldatacopy(0, 0, 0, 0, x1, x2, x3, x4, 0, 0, 0, 32:i64)
+	z1, z2, z3, z4 := mload_internal(0:i32)
+}
+
+// size of call data in bytes
+function calldatasize() -> z1, z2, z3, z4 {
+	z4 := i64.extend_i32_u(eth.getCallDataSize())
+}
+
+// copy z bytes from calldata at position y to mem at position x
+function calldatacopy(x1, x2, x3, x4, y1, y2, y3, y4, z1, z2, z3, z4) {
+	let cds:i32 := eth.getCallDataSize()
+	let offset:i32 := u256_to_i32(y1, y2, y3, y4)
+	let size:i32 := u256_to_i32(z1, z2, z3, z4)
+	// overflowed?
+	if i32.gt_u(offset, i32.sub(0xffffffff:i32, size)) {
+		eth.revert(0:i32, 0:i32)
+	}
+	if i32.gt_u(i32.add(size, offset), cds) {
+		size := i32.sub(cds, offset)
+	}
+	if i32.gt_u(size, 0:i32) {
+		eth.callDataCopy(
+			u256_to_i32(x1, x2, x3, x4),
+			offset,
+			size
+		)
+	}
+}
+
+// Needed?
+// size of call data in bytes
+function codesize() -> z1, z2, z3, z4 {
+	z4 := i64.extend_i32_u(eth.getCodeSize())
+}
+
+// copy z bytes from code at position y to mem at position x
+function codecopy(x1, x2, x3, x4, y1, y2, y3, y4, z1, z2, z3, z4) {
+	eth.codeCopy(
+		to_internal_i32ptr(x1, x2, x3, x4),
+		u256_to_i32(y1, y2, y3, y4),
+		u256_to_i32(z1, z2, z3, z4)
+	)
+}
+
+//
+function datacopy(x1, x2, x3, x4, y1, y2, y3, y4, z1, z2, z3, z4) {
+	// TODO correct?
+	codecopy(x1, x2, x3, x4, y1, y2, y3, y4, z1, z2, z3, z4)
+}
+
+// gas price of the transaction
+function gasprice() -> z1, z2, z3, z4 {
+	eth.getTxGasPrice(0:i32)
+	z1, z2, z3, z4 := mload_internal(0:i32)
+}
+
+// size of the code at address x, returns i32
+function extcodesize_internal(x1, x2, x3, x4) -> r:i32 {
+	mstore_address(0:i32, x1, x2, x3, x4)
+	r := eth.getExternalCodeSize(12:i32)
+}
+
+// size of the code at address x, returns u256
+function extcodesize(x1, x2, x3, x4) -> z1, z2, z3, z4 {
+	z4 := i64.extend_i32_u(extcodesize_internal(x1, x2, x3, x4))
+}
+
+// code hash of address x
+function extcodehash(x1, x2, x3, x4) -> z1, z2, z3, z4 {
+	// TODO: not part of current Ewasm spec
+	unreachable()
+}
+
+// like codecopy(t, f, s) but take code at address a
+function extcodecopy(a1, a2, a3, a4, p1, p2, p3, p4, o1, o2, o3, o4, l1, l2, l3, l4) {
+	mstore_address(0:i32, a1, a2, a3, a4)
+	let codeOffset:i32 := u256_to_i32(o1, o2, o3, o4)
+	let codeLength:i32 := u256_to_i32(l1, l2, l3, l4)
+	eth.externalCodeCopy(12:i32, to_internal_i32ptr(p1, p2, p3, p4), codeOffset, codeLength)
+}
+
+// size of the last returndata
+function returndatasize() -> z1, z2, z3, z4 {
+	z4 := i64.extend_i32_u(eth.getReturnDataSize())
+}
+
+// returndatacopy(t, f, s): copy s bytes from returndata at position f to mem at position t
+function returndatacopy(x1, x2, x3, x4, y1, y2, y3, y4, z1, z2, z3, z4) {
+	eth.returnDataCopy(
+		to_internal_i32ptr(x1, x2, x3, x4),
+		u256_to_i32(y1, y2, y3, y4),
+		u256_to_i32(z1, z2, z3, z4)
+	)
+}
+
+// blockhash(b): hash of block nr b - only for last 256 blocks excluding current
+function blockhash(x1, x2, x3, x4) -> z1, z2, z3, z4 {
+	let r:i32 := eth.getBlockHash(u256_to_i64(x1, x2, x3, x4), 0:i32)
+	if i32.eqz(r) {
+		z1, z2, z3, z4 := mload_internal(0:i32)
+	}
+}
+
+// coinbase(): current mining beneficiary
+function coinbase() -> z1, z2, z3, z4 {
+	eth.getBlockCoinbase(4:i32)
+	z2, z3, z4 := mload_address(0:i32)
+}
+
+// timestamp(): timestamp of the current block in seconds since the epoch
+function timestamp() -> z1, z2, z3, z4 {
+	z4 := eth.getBlockTimestamp()
+}
+
+// number(): current block number
+function number() -> z1, z2, z3, z4 {
+	z4 := eth.getBlockNumber()
+}
+
+// difficulty(): difficulty of the current block
+function difficulty() -> z1, z2, z3, z4 {
+	eth.getBlockDifficulty(0:i32)
+	z1, z2, z3, z4 := mload_internal(0:i32)
+}
+
+// gaslimit(): block gas limit of the current block
+function gaslimit() -> z1, z2, z3, z4 {
+	z4 := eth.getBlockGasLimit()
+}
+
+// mload(p): mem[p…(p+32))
+function mload(x1, x2, x3, x4) -> z1, z2, z3, z4 {
+	z1, z2, z3, z4 := mload_internal(to_internal_i32ptr(x1, x2, x3, x4))
+}
+
+// mload_internal(p): bswap(mem[p…(p+32)))
+function mload_internal(pos:i32) -> z1, z2, z3, z4 {
+	z1 := bswap64(i64.load(pos))
+	z2 := bswap64(i64.load(i32.add(pos, 8:i32)))
+	z3 := bswap64(i64.load(i32.add(pos, 16:i32)))
+	z4 := bswap64(i64.load(i32.add(pos, 24:i32)))
+}
+
+// mload_address(pos:i32): loads address from pos
+function mload_address(pos:i32) -> z2, z3, z4 {
+	z2 := i64.extend_i32_u(bswap32(i32.load(i32.add(pos, 4:i32))))
+	z3 := bswap64(i64.load(i32.add(pos, 8:i32)))
+	z4 := bswap64(i64.load(i32.add(pos, 16:i32)))
+}
+
+// mstore(p, v): mem[p…(p+32)) := v
+function mstore(x1, x2, x3, x4, y1, y2, y3, y4) {
+	mstore_internal(to_internal_i32ptr(x1, x2, x3, x4), y1, y2, y3, y4)
+}
+
+// mstore(p, v): mem[p…(p+32)) := bswap(v)
+function mstore_internal(pos:i32, y1, y2, y3, y4) {
+	i64.store(pos, bswap64(y1))
+	i64.store(i32.add(pos, 8:i32), bswap64(y2))
+	i64.store(i32.add(pos, 16:i32), bswap64(y3))
+	i64.store(i32.add(pos, 24:i32), bswap64(y4))
+}
+
+// mstore_address(p, a): writes bswap(address) a to pos
+function mstore_address(pos:i32, a1, a2, a3, a4) {
+	a1, a2, a3 := u256_to_address(a1, a2, a3, a4)
+	mstore_internal(pos, 0, a1, a2, a3)
+}
+
+// mstore8(p, v): mem[p] := v & 0xff (only modifies a single byte)
+function mstore8(x1, x2, x3, x4, y1, y2, y3, y4) {
+	let v := u256_to_byte(y1, y2, y3, y4)
+	i64.store8(to_internal_i32ptr(x1, x2, x3, x4), v)
+}
+
+// Needed?
+// size of memory, i.e. largest accessed memory index
+function msize() -> z1, z2, z3, z4 {
+	// TODO implement
+	unreachable()
+}
+
+// sload(p): storage[p]
+function sload(x1, x2, x3, x4) -> z1, z2, z3, z4 {
+	mstore_internal(0:i32, x1, x2, x3, x4)
+	eth.storageLoad(0:i32, 32:i32)
+	z1, z2, z3, z4 := mload_internal(32:i32)
+}
+
+// sstore(p, v): storage[p] := v
+function sstore(x1, x2, x3, x4, y1, y2, y3, y4) {
+	mstore_internal(0:i32, x1, x2, x3, x4)
+	mstore_internal(32:i32, y1, y2, y3, y4)
+	eth.storageStore(0:i32, 32:i32)
+}
+
+// gas(): gas still available to execution
+function gas() -> z1, z2, z3, z4 {
+	z4 := eth.getGasLeft()
+}
+
+// log0(p, s): log without topics and data mem[p…(p+s))
+function log0(p1, p2, p3, p4, s1, s2, s3, s4) {
+	eth.log(
+		to_internal_i32ptr(p1, p2, p3, p4),
+		u256_to_i32(s1, s2, s3, s4),
+		0:i32, 0:i32, 0:i32, 0:i32, 0:i32
+	)
+}
+
+// log1(p, s, t1): log with topic t1 and data mem[p…(p+s))
+function log1(
+	p1, p2, p3, p4, s1, s2, s3, s4,
+	t1_1, t1_2, t1_3, t1_4
+) {
+	eth.log(
+		to_internal_i32ptr(p1, p2, p3, p4),
+		u256_to_i32(s1, s2, s3, s4),
+		1:i32,
+		to_internal_i32ptr(t1_1, t1_2, t1_3, t1_4),
+		0:i32, 0:i32, 0:i32
+	)
+}
+
+// log2(p, s, t1, t2): log with topics t1, t2 and data mem[p…(p+s))
+function log2(
+	p1, p2, p3, p4, s1, s2, s3, s4,
+	t1_1, t1_2, t1_3, t1_4,
+	t2_1, t2_2, t2_3, t2_4
+) {
+	eth.log(
+		to_internal_i32ptr(p1, p2, p3, p4),
+		u256_to_i32(s1, s2, s3, s4),
+		2:i32,
+		to_internal_i32ptr(t1_1, t1_2, t1_3, t1_4),
+		to_internal_i32ptr(t2_1, t2_2, t2_3, t2_4),
+		0:i32, 0:i32
+	)
+}
+
+// log3(p, s, t1, t2, t3): log with topics t1, t2, t3 and data mem[p…(p+s))
+function log3(
+	p1, p2, p3, p4, s1, s2, s3, s4,
+	t1_1, t1_2, t1_3, t1_4,
+	t2_1, t2_2, t2_3, t2_4,
+	t3_1, t3_2, t3_3, t3_4
+) {
+	eth.log(
+		to_internal_i32ptr(p1, p2, p3, p4),
+		u256_to_i32(s1, s2, s3, s4),
+		3:i32,
+		to_internal_i32ptr(t1_1, t1_2, t1_3, t1_4),
+		to_internal_i32ptr(t2_1, t2_2, t2_3, t2_4),
+		to_internal_i32ptr(t3_1, t3_2, t3_3, t3_4),
+		0:i32
+	)
+}
+
+// log4(p, s, t1, t2, t3, t4): log with topics t1, t2, t3, t4 and data mem[p…(p+s))
+function log4(
+	p1, p2, p3, p4, s1, s2, s3, s4,
+	t1_1, t1_2, t1_3, t1_4,
+	t2_1, t2_2, t2_3, t2_4,
+	t3_1, t3_2, t3_3, t3_4,
+	t4_1, t4_2, t4_3, t4_4,
+) {
+	eth.log(
+		to_internal_i32ptr(p1, p2, p3, p4),
+		u256_to_i32(s1, s2, s3, s4),
+		4:i32,
+		to_internal_i32ptr(t1_1, t1_2, t1_3, t1_4),
+		to_internal_i32ptr(t2_1, t2_2, t2_3, t2_4),
+		to_internal_i32ptr(t3_1, t3_2, t3_3, t3_4),
+		to_internal_i32ptr(t4_1, t4_2, t4_3, t4_4)
+	)
+}
+
+// create(v, p, n): create new contract with code mem[p…(p+n)) and send v wei
+//     and return the new address.
+function create(
+	x1, x2, x3, x4,
+	y1, y2, y3, y4,
+	z1, z2, z3, z4
+) -> a1, a2, a3, a4 {
+	let v1, v2 := u256_to_u128(x1, x2, x3, x4)
+	mstore_internal(0:i32, 0, 0, v1, v2)
+
+	let r:i32 := eth.create(0:i32, to_internal_i32ptr(y1, y2, y3, y4), u256_to_i32(z1, z2, z3, z4), 32:i32)
+	if i32.eqz(r) {
+		a1, a2, a3, a4 := mload_internal(32:i32)
+	}
+}
+
+// call(g, a, v, in, insize, out, outsize): call contract at address a with
+//     input mem[in…(in+insize)) providing g gas and v wei and output area
+//     mem[out…(out+outsize)) returning 0 on error (eg. out of gas) and 1 on success.
+function call(
+	a1, a2, a3, a4,
+	b1, b2, b3, b4,
+	c1, c2, c3, c4,
+	d1, d2, d3, d4,
+	e1, e2, e3, e4,
+	f1, f2, f3, f4,
+	g1, g2, g3, g4
+) -> x1, x2, x3, x4 {
+	let g := u256_to_i64(a1, a2, a3, a4)
+	mstore_address(0:i32, b1, b2, b3, b4)
+
+	let v1, v2 := u256_to_u128(c1, c2, c3, c4)
+	mstore_internal(32:i32, 0, 0, v1, v2)
+
+	x4 := i64.extend_i32_u(eth.call(g, 12:i32, 32:i32, to_internal_i32ptr(d1, d2, d3, d4), u256_to_i32(e1, e2, e3, e4)))
+}
+
+// callcode(g, a, v, in, insize, out, outsize): identical to call but only use
+//     the code from a and stay in the context of the current contract otherwise.
+function callcode(
+	a1, a2, a3, a4,
+	b1, b2, b3, b4,
+	c1, c2, c3, c4,
+	d1, d2, d3, d4,
+	e1, e2, e3, e4,
+	f1, f2, f3, f4,
+	g1, g2, g3, g4
+) -> x1, x2, x3, x4 {
+	mstore_address(0:i32, b1, b2, b3, b4)
+
+	let v1, v2 := u256_to_u128(c1, c2, c3, c4)
+	mstore_internal(32:i32, 0, 0, v1, v2)
+
+	x4 := i64.extend_i32_u(eth.callCode(
+		u256_to_i64(a1, a2, a3, a4),
+		12:i32,
+		32:i32,
+		to_internal_i32ptr(d1, d2, d3, d4),
+		u256_to_i32(e1, e2, e3, e4)
+	))
+}
+
+// delegatecall(g, a, in, insize, out, outsize): identical to callcode
+//     but also keep caller and callvalue.
+function delegatecall(
+	a1, a2, a3, a4,
+	b1, b2, b3, b4,
+	c1, c2, c3, c4,
+	d1, d2, d3, d4,
+	e1, e2, e3, e4,
+	f1, f2, f3, f4
+) -> x1, x2, x3, x4 {
+	mstore_address(0:i32, b1, b2, b3, b4)
+
+	x4 := i64.extend_i32_u(eth.callDelegate(
+		u256_to_i64(a1, a2, a3, a4),
+		12:i32,
+		to_internal_i32ptr(c1, c2, c3, c4),
+		u256_to_i32(d1, d2, d3, d4)
+	))
+}
+
+// staticcall(g, a, in, insize, out, outsize): identical to
+//     call(g, a, 0, in, insize, out, outsize) but do not allow state modifications.
+function staticcall(
+	a1, a2, a3, a4,
+	b1, b2, b3, b4,
+	c1, c2, c3, c4,
+	d1, d2, d3, d4,
+	e1, e2, e3, e4,
+	f1, f2, f3, f4
+) -> x1, x2, x3, x4 {
+	mstore_address(0:i32, b1, b2, b3, b4)
+
+	x4 := i64.extend_i32_u(eth.callStatic(
+		u256_to_i64(a1, a2, a3, a4),
+		12:i32,
+		to_internal_i32ptr(c1, c2, c3, c4),
+		u256_to_i32(d1, d2, d3, d4)
+	))
+}
+
+// create2(v, p, n, s): create new contract with code mem[p…(p+n)) at
+//     address keccak256(0xff . this . s . keccak256(mem[p…(p+n))) and
+//     send v wei and return the new address, where 0xff is a 1 byte value,
+//     this is the current contract’s address as a 20 byte value and s
+//     is a big-endian 256-bit value.
+function create2(
+	a1, a2, a3, a4,
+	b1, b2, b3, b4,
+	c1, c2, c3, c4,
+	d1, d2, d3, d4
+) -> x1, x2, x3, x4 {
+	// TODO: not part of current Ewasm spec
+	unreachable()
+}
+
+// selfdestruct(a): end execution, destroy current contract and send funds to a
+function selfdestruct(a1, a2, a3, a4) {
+	mstore_address(0:i32, a1, a2, a3, a4)
+	// In EVM, addresses are padded to 32 bytes, so discard the first 12.
+	eth.selfDestruct(12:i32)
+}
+
+// return(p, s): end execution, return data mem[p…(p+s)).
+function return(x1, x2, x3, x4, y1, y2, y3, y4) {
+	eth.finish(
+		to_internal_i32ptr(x1, x2, x3, x4),
+		u256_to_i32(y1, y2, y3, y4)
+	)
+}
+
+// revert(p, s): end execution, revert state changes, return data mem[p…(p+s)).
+function revert(x1, x2, x3, x4, y1, y2, y3, y4) {
+	eth.revert(
+		to_internal_i32ptr(x1, x2, x3, x4),
+		u256_to_i32(y1, y2, y3, y4)
+	)
+}
+
+// invalid(): end execution with invalid instruction.
+function invalid() {
+	unreachable()
+}
+
+// stop(): stop execution, identical to return(0, 0).
+function stop() {
+	eth.finish(0:i32, 0:i32)
+}
+)"};
+
+} // namespace solidity::yul::wasm::polyfill
