+++ libyul/backends/wasm/polyfill/keccak.h
@@ -1,0 +1,39 @@
+/*
+	This file is part of solidity.
+
+	solidity is free software: you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation, either version 3 of the License, or
+	(at your option) any later version.
+
+	solidity is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with solidity.  If not, see <http://www.gnu.org/licenses/>.
+*/
+// SPDX-License-Identifier: GPL-3.0
+/**
+ * This file includes logical functions for the wasm polyfill.
+ */
+
+#pragma once
+
+#include <string>
+
+namespace solidity::yul::wasm::polyfill
+{
+
+static std::string const keccak{R"(
+// solidity::yul::wasm::polyfill::keccak
+
+// keccak256(p, n): keccak(mem[p…(p+n)))
+function keccak256(x1, x2, x3, x4, y1, y2, y3, y4) -> z1, z2, z3, z4 {
+	// TODO implement
+	unreachable()
+}
+)"};
+
+} // namespace solidity::yul::wasm::polyfill
